- Kill the last compiler warning under 5.3 (release-only)
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_3@261397 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/phar/tar.c
@@ -1087,7 +1087,7 @@ nostub:
 #endif
 		PHAR_SET_32(sigbuf, phar->sig_flags);
 		PHAR_SET_32(sigbuf + 4, signature_length);
-		if (8 != php_stream_write(entry.fp, sigbuf, 8) || signature_length != php_stream_write(entry.fp, signature, signature_length)) {
+		if (8 != (int)php_stream_write(entry.fp, sigbuf, 8) || signature_length != (int)php_stream_write(entry.fp, signature, signature_length)) {
 			efree(signature);
 			if (error) {
 				spprintf(error, 0, "phar error: unable to write signature to tar-based phar %s", phar->fname);
